requirements/local.txt
-r common.txt

faker==19.2.0
model-bakery==1.13.0
pytest==7.4.0
pytest-django==4.5.2
